TODO/requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
drf-yasg==1.21.3
Faker==12.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mixer==7.2.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
promise==2.3
psycopg2==2.9.3
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-multipart==0.0.5
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
